Assets added, initial project setup
